+++ plugins/qcheck-stm/test/module_prefix.ml
@@ -1,0 +1,3 @@
+type 'a t = 'a ref
+
+let make a = ref a